+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2018-08-09
+### Added
+- Added an example app comparing `UICollectionViewFlowLayout` and `FlexibleRowHeightGridLayout`.
+
+### Changed
+- Fixed issue whereby `FlexibleRowHeightGridLayout` initializer was inaccessible due not being marked as public.
+- In `FlexibleRowHeightGridLayoutDelegate` the following methods have been changed:
+	- `func collectionView(_ collectionView: UICollectionView, heightForItemAt indexPath: IndexPath) -> CGFloat` *becomes ->* `@objc optional func collectionView(_ collectionView: UICollectionView, layout: FlexibleRowHeightGridLayout, referenceHeightForFooterInSection section: Int) -> CGFloat`
+	- `@objc optional func collectionView(_ collectionView: UICollectionView, referenceHeightForHeaderInSection section: Int) -> CGFloat` *becomes ->* `@objc optional func collectionView(_ collectionView: UICollectionView, layout: FlexibleRowHeightGridLayout, referenceHeightForFooterInSection section: Int) -> CGFloat`
+	- `@objc optional func collectionView(_ collectionView: UICollectionView, referenceHeightForFooterInSection section: Int) -> CGFloat` *becomes ->* `@objc optional func collectionView(_ collectionView: UICollectionView, layout: FlexibleRowHeightGridLayout, referenceHeightForFooterInSection section: Int) -> CGFloat`
+
 ## [1.0.0] - 2018-08-08
 ### Added
 - Support for sections including headers and / or footers.